Refactoring FileOutput
FileOutput does not assume PetscViewer anymore, since some file outputs
are just not using it.  This created minor code duplication in VTKOutput
and HDF5Output which are PetscViewer-based.  But, VTKOutput is mainly
just for trivial testing and HDF5Output is not being used that much
either (which may change in future) - so this is not a big problem.

Anyway, if the code duplication is a burning problem, we can add an
intermediate class and move the common code there.