Renaming `set_up_aux_fields` to `create_aux_fields`
There is a method called `set_up_fields` which is meant to set both primary
and auxiliary fields.  Thus, there could be a confusion about what
`set_up_aux_fields` was for. This method is really for internal use and should
not be used by the application code. Thus, the rename.