Allow iteration over empty `IndexSet`
Now, we can run a range-based for-loop over an empty `IndexSet` without
crashing. This is useful for cases where for example domain partitioning
produces an empty index set. Now, we do not have to check if the index set
is empty or not. We simply run an empty for loop. #quality-of-life