`pre_solve` and `post_solve` are virtual
I finally convinced myself that `pre_solve` and `post_solve` should be virtual.
This allows user applications to simply insert their own code and retain the coding
in `run` which is called from the main app. This is the case for just standalone apps.
Coupled apps will be able to still call `pre_solve` and `post_solve` when needed, since
these are still public.

Note that we also have `pre_step` and `post_step` which are also virtual.