Do not register custom MPI datatype for DenseVector<T, 1>
We need to use `mpi_datatype` for DenseVector<T, 1> rather then
`type_contiguous(1, mpi_datatype<T>())`, which seems to fubar MPI
internally and it starts to see things like `double` as the custom
defined type and that's when things break interestingly.

We prevent the code from compiling when somebody tries to use
DenseVector<T, 1>. We use `mpi_datatype` for DenseVector<Int, 1>
and DenseVector<Real, 1> which covers our current needs.